% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{as.matrix.foodweb}
\alias{as.matrix.foodweb}
\title{Convert a \code{foodweb} to a matrix}
\usage{
\method{as.matrix}{foodweb}(x, rownames.force = NA, ...)
}
\arguments{
\item{x}{A \code{foodweb}}

\item{rownames.force, ...}{Ignored, only included for compatibility with S3 generic}
}
\description{
This is equivalent to calling \code{\link[=get_funmat]{get_funmat()}} on \code{x}.
}
\keyword{internal}
