% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{as_tbl_graph.foodweb}
\alias{as_tbl_graph.foodweb}
\title{Convert a \code{foodweb} to a \code{tidygraph}}
\usage{
\method{as_tbl_graph}{foodweb}(x, ...)
}
\arguments{
\item{x}{A \code{foodweb} object created by \code{\link[=foodweb]{foodweb()}}.}
}
\value{
A new \link[tidygraph:tbl_graph]{tidygraph::tbl_graph} object.
}
\description{
This is an S3 method for the generic \code{\link[tidygraph:as_tbl_graph]{tidygraph::as_tbl_graph()}}.
}
\keyword{internal}
