% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_namespaced_function_name}
\alias{build_namespaced_function_name}
\title{Standardise function names}
\usage{
build_namespaced_function_name(fn, ns)
}
\arguments{
\item{fn}{Character vector of function names}

\item{ns}{Character vector of namespaces}
}
\value{
Character vector of fully-qualified function names
}
\description{
Given a function name that may or may not contain a namespace (\code{"foo::bar"} or \code{"bar"}) and the
namespace to which that function belongs (\code{"foo"}), return \code{"foo::bar"}.
}
\keyword{internal}
