% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expr_is_bare_function}
\alias{expr_is_bare_function}
\title{Is an expression a function?}
\usage{
expr_is_bare_function(expr, env = parent.frame())
}
\arguments{
\item{expr}{An expression}

\item{env}{The environment to look in}
}
\value{
Boolean
}
\description{
E.g. \code{mean} or \code{cowsay::say} (not invoked)
}
\keyword{internal}
