% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-matrix.R
\name{foodweb_matrix}
\alias{foodweb_matrix}
\title{Create a function caller/callee matrix}
\usage{
foodweb_matrix(env = parent.frame())
}
\arguments{
\item{env}{Environment in which to search for functions.}
}
\value{
An n x n matrix where \emph{n} is the number of functions in \code{env}.
}
\description{
Returns a matrix of 0s and 1s with a row and column for each function in an environment, such
that if the function on the x-axis calls the function on the y-axis, the element is 1, otherwise
0.
}
