% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{foodweb_summarise}
\alias{foodweb_summarise}
\title{Utility function to pluralise "edge" or "edges"}
\usage{
foodweb_summarise(x)
}
\arguments{
\item{x}{A \code{foodweb}}
}
\value{
A character string in the format "x nodes and y edge/s"
}
\description{
Used when generating descriptions of a foodweb.
}
\keyword{internal}
