% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-matrix.R
\name{functions_called_by}
\alias{functions_called_by}
\title{Which functions does a function call?}
\usage{
functions_called_by(fn_name, funs_to_match, env)
}
\arguments{
\item{fn_name}{\verb{<chr>} The name of the function of interest.}

\item{funs_to_match}{\verb{<chr>} Only these functions will be considered as callees.}

\item{env}{\verb{<env>} The environment in which \code{fn_name} lives.}
}
\value{
An integer vector of positions in \code{funs_to_match}.
}
\description{
Given an input function \code{fn_name} and a list of candidate functions \code{funs_to_match}, return the
indices in \code{funs_to_match} of functions that \code{fn_name} calls.
}
\details{
As well as direct calls like \code{foo()}, this detects functions passed to \code{\link[=do.call]{do.call()}}, the
\code{lapply()} family and purrr, e.g. \code{do.call("foo", args)}, \code{lapply(x, foo)} or
\code{purrr::map(x, ~ foo(.x))}. See \code{\link[=functions_passed_by]{functions_passed_by()}}.
}
\keyword{internal}
