% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_called_functions}
\alias{get_called_functions}
\title{Return the names of all the functions called by another}
\usage{
get_called_functions(FUN)
}
\arguments{
\item{FUN}{A function}
}
\value{
A character vector containing names of all (non-primitive) functions called by \code{FUN}
}
\description{
Return the names of all the functions called by another
}
\keyword{internal}
