% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_calls}
\alias{get_calls}
\title{Extract all call objects from a function body}
\usage{
get_calls(expr)
}
\arguments{
\item{expr}{The result of calling \code{\link[=body]{body()}} on a function.}
}
\value{
A nested list of function symbols found in \code{expr}.
}
\description{
Given an input function body, return all the (non-primitive) calls to other functions within it.
The returned data structure is a nested list reflecting the structure of the source code.
}
\keyword{internal}
