% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{get_funmat}
\alias{get_funmat}
\title{Extract the function matrix from a \code{foodweb} object.}
\usage{
get_funmat(x)
}
\arguments{
\item{x}{A \code{foodweb}}
}
\value{
\code{x$funmat} - a numeric matrix.
}
\description{
Extract the function matrix from a \code{foodweb} object.
}
