% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{get_graphviz_spec}
\alias{get_graphviz_spec}
\title{Extract the GraphViz specification from a \code{foodweb} object.}
\usage{
get_graphviz_spec(x)
}
\arguments{
\item{x}{A \code{foodweb}}
}
\value{
\code{x$graphviz_spec} - a character scalar.
}
\description{
Extract the GraphViz specification from a \code{foodweb} object.
}
