% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{graph_spec_from_matrix}
\alias{graph_spec_from_matrix}
\title{Create a \code{grViz} specification from a function matrix}
\usage{
graph_spec_from_matrix(funmat, title = "foodweb")
}
\arguments{
\item{funmat}{A function matrix generated by \code{\link[=function_matrix]{function_matrix()}}.}

\item{title}{The name of the graph. Will be displayed as mouseover text.}
}
\value{
A text string.
}
\description{
Given a function matrix created by \code{\link[=function_matrix]{function_matrix()}}, convert it into a text specification
that can be passed to \code{\link[DiagrammeR:grViz]{DiagrammeR::grViz()}}.
}
