% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{new_foodweb}
\alias{new_foodweb}
\title{Create a new \code{foodweb} object}
\usage{
new_foodweb(funmat, title)
}
\arguments{
\item{funmat}{A function matrix created by \code{\link[=function_matrix]{function_matrix()}}}

\item{title}{Title to appear on the default graph visualisation}
}
\value{
A \code{foodweb}.
}
\description{
A \code{foodweb} object describes the relationship of functions in an environment. It has two
components: a \code{funmat} (function matrix) which encodes the caller/callee relationships (i.e.
which functions call which) and a \code{grviz_spec} which is a text representation of the graph and
is used for the default plotting behaviour.
}
\details{
This function should not be called directly, use \code{\link[=foodweb]{foodweb()}} instead.
}
\seealso{
foodweb
}
\keyword{internal}
