% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{plot.foodweb}
\alias{plot.foodweb}
\title{Plot a \code{foodweb} object}
\usage{
\method{plot}{foodweb}(x, ...)
}
\arguments{
\item{x}{A \code{foodweb} object.}

\item{...}{Unused, only included for consistency with S3 generic.}
}
\description{
Calls \code{\link[DiagrammeR:grViz]{DiagrammeR::grViz()}} on the \code{graphvis_spec} element of the \code{foodweb}.
}
