% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{print.foodweb}
\alias{print.foodweb}
\title{Print a \code{foodweb} object}
\usage{
\method{print}{foodweb}(x, ...)
}
\arguments{
\item{x}{A \code{foodweb} object.}

\item{...}{Unused, only included for consistency with S3 generic.}
}
\description{
Prints the \code{graphvis_spec} member of a \code{foodweb} object.
}
\keyword{internal}
