% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-matrix.R
\name{print.foodweb_matrix}
\alias{print.foodweb_matrix}
\title{Print a \code{foodweb_matrix}}
\usage{
\method{print}{foodweb_matrix}(x, ...)
}
\arguments{
\item{x}{A \code{foodweb_matrix}}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a \code{foodweb_matrix}
}
