% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{tokenise_function}
\alias{tokenise_function}
\title{Convert a function (a language object) to text tokenstype}
\usage{
tokenise_function(x)
}
\description{
Convert a function (a language object) to text tokenstype
}
