% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Filter a function matrix}
\usage{
filter_matrix(fn_name, fn_mat)
}
\arguments{
\item{fn_name}{String giving the name of the function we're interested in}

\item{fn_mat}{Matrix prpduced by \code{\link[=function_matrix]{function_matrix()}}}
}
\value{
A filtered function matrix containing only functions that are direct descendants or
antecedants of \code{fn_name}.
}
\description{
Filter a function matrix
}
\keyword{internal}
