% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{foodweb}
\alias{foodweb}
\title{Create a foodweb}
\usage{
foodweb(FUN = NULL, env = .GlobalEnv, as.text = FALSE)
}
\arguments{
\item{FUN}{A function}

\item{env}{An environment, \code{.GlobalEnv} by default. Ignored if \code{FUN} is not \code{NULL}}

\item{as.text}{Boolean. If \code{TRUE}, rather than rendering the graph the intermediate graphviz
specification is returned.}
}
\value{
If \code{as.text} is \code{TRUE}, a character vector. Otherwise nothing is returned and the
function only renders the graph.
}
\description{
Create a foodweb
}
\examples{
foodweb()

foodweb(as.text = TRUE)

# "cowsay" is a small enough package that this output is readable. Try this with larger packages
# at your peril.
if (requireNamespace("cowsay", quietly = TRUE)) {
  foodweb(FUN = cowsay::say)
}
}
