% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{functions_called_by}
\alias{functions_called_by}
\title{Which functions does a function call?}
\usage{
functions_called_by(fn_name, funs_to_match, where)
}
\arguments{
\item{fn_name}{\verb{<chr>}\cr The name of the function of interest}

\item{funs_to_match}{\verb{<chr>}\cr Only these functions will be considered as parents}

\item{where}{\verb{<env>}\cr An environment, or text specifying an environment}
}
\value{
A character vector listing the functions in \code{funs_to_match} that call \code{fn_name}.
}
\description{
Given an input function \code{fn_name} and a list of candidate functions \code{funs_to_match}, return a
list of all the functions in \code{funs_to_match} that appear in the definition of \code{fn_name}.
}
\keyword{internal}
