% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{new_foodweb}
\alias{new_foodweb}
\title{Create a new \code{foodweb} object}
\usage{
new_foodweb(funmat, title)
}
\arguments{
\item{funmat}{A function matrix created by \code{\link[=function_matrix]{function_matrix()}}}

\item{title}{Title to appear on the default graph visualisation}
}
\value{
A \code{foodweb}.
}
\description{
Create a new \code{foodweb} object
}
