% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodweb.R
\name{plot.foodweb}
\alias{plot.foodweb}
\title{Plot a \code{foodweb} object}
\usage{
\method{plot}{foodweb}(x, ...)
}
\arguments{
\item{x}{A \code{foodweb} object.}

\item{...}{Further arguments to be passed to \code{\link[DiagrammeR:grViz]{DiagrammeR::grViz}}}
}
\value{
An object of class \code{htmlwidget}; the output is just that of the
underlying call to \code{\link[DiagrammeR:grViz]{DiagrammeR::grViz}}.
}
\description{
Calls \code{\link[DiagrammeR:grViz]{DiagrammeR::grViz()}} on the \code{graphvis_spec} element of the \code{foodweb}.
}
\keyword{internal}
