% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebr.R
\name{print.foodweb}
\alias{print.foodweb}
\title{Print a \code{foodweb} object}
\usage{
\method{print}{foodweb}(x)
}
\arguments{
\item{x}{A \code{foodweb} object.}
}
\description{
Prints the \code{grvis_spec} member of a \code{foodweb} object.
}
