% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-matrix.R
\name{tokenise_function}
\alias{tokenise_function}
\title{Convert a function body to text}
\usage{
tokenise_function(x)
}
\arguments{
\item{x}{A function}
}
\value{
A character string containing the tokenised function body
}
\description{
Convert a function body to text
}
\keyword{internal}
